option.useFileName
Option to use the filename as the variable name. If option.useFileName is a string, the variable name will have that string removed from the file name. E.g. a useFileName value of '../public_html/' will change an absolute path of '../public_html/img/photo.jpg' to 'img/photo.jpg'